# Optimise getting local members and membership counts